feature(Refactor): Move ResourceManager to resource package
- Move ResourceManager to resource
- Move ProgressLIstener to resource
- Fix some discrepancies
