+++ src/main/resources/res/assets/asset-2.json
@@ -69,7 +69,7 @@
 			"layer": 2,
 			"color": "[0.6,0.5,0.12,0.98]",
 			"attributes": {
-				"intensity": 0.3
+				"intensity": 0.1
 				}
 		},
 		"l": {
@@ -82,7 +82,7 @@
 			"layer": 4,
 			"color": "[0.9,0.9,0.3,1.0]",
 			"attributes": {
-				"intensity": 0.6,
+				"intensity": 0.3,
 				"radius": 130,
 				"glittering": 0.4
 			}
